requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
